+++ src/sys/os/linux/syscalls/x86.rs
@@ -9,7 +9,7 @@ use super::{LinuxOffset, shared_docs::*};
 use crate::{LINUX_SYS as SYS, Linux, LinuxSigaction, LinuxStat};
 #[cfg(feature = "time")]
 use crate::{LinuxClock, LinuxTimespec};
-use crate::{asm, c_char, c_int, c_uchar, c_uint, c_ulong};
+use crate::{asm, c_char, c_int, c_mode_t, c_uchar, c_uint, c_ulong};
 
 /// # Syscalls: File descriptors.
 impl Linux {
@@ -62,6 +62,32 @@ impl Linux {
                 in("edx") mode,
                 lateout("eax") result,
                 options(nostack)
+            );
+        }
+        result
+    }
+    #[must_use]
+    #[doc = _DOC_SYS_OPENAT!()]
+    pub unsafe fn sys_openat(
+        dirfd: c_int,
+        path: *const c_char,
+        flags: c_int,
+        mode: c_mode_t,
+    ) -> c_int {
+        let result: c_int;
+        unsafe {
+            asm!(
+                "push esi",
+                "mov esi, {mode:e}",
+                "mov eax, {OPENAT}",
+                "int 0x80",
+                "pop esi",
+                OPENAT = const SYS::OPENAT,
+                mode = in(reg) mode,
+                in("ebx") dirfd,
+                in("ecx") path,
+                in("edx") flags,
+                lateout("eax") result,
             );
         }
         result